+++ packages/devices/src/index.js
@@ -11,14 +11,7 @@ module.exports = ({
   ]),
   lossyDeviceName = false
 } = {}) => {
-  const devices = puppeteerDevices.concat(customDevices).reduce(
-    (acc, { name, ...props }) => ({
-      ...acc,
-      [name]: props
-    }),
-    {}
-  )
-
+  const devices = { ...puppeteerDevices.devicesMap, ...customDevices }
   const deviceDescriptors = Object.keys(devices)
 
   const findDevice = (deviceDescriptor, lossyEnabled) => {
@@ -36,13 +29,13 @@ module.exports = ({
 
     return device
       ? {
-        userAgent: device.userAgent || headers['user-agent'],
-        viewport: { ...device.viewport, ...viewport }
-      }
+          userAgent: device.userAgent || headers['user-agent'],
+          viewport: { ...device.viewport, ...viewport }
+        }
       : {
-        userAgent: headers['user-agent'],
-        viewport
-      }
+          userAgent: headers['user-agent'],
+          viewport
+        }
   }
 
   getDevices.devices = devices