+++ packages/screenshot/src/index.js
@@ -1,16 +1,22 @@
 'use strict'
 
 const debug = require('debug-logfmt')('browserless:screenshot')
+
 const { extension } = require('mime-types')
 const prettyMs = require('pretty-ms')
 const timeSpan = require('time-span')
 const pReflect = require('p-reflect')
 
+const pretty = require('./pretty')
 const createGoto = require('./goto')
 const overlay = require('./overlay')
-const pretty = require('./pretty')
 
-const isJSON = headers => extension(headers['content-type']) === 'json'
+const PRETTY_CONTENT_TYPES = ['json', 'text', 'html']
+
+const getContentType = headers => {
+  const contentType = extension(headers['content-type'])
+  return contentType === 'txt' ? 'text' : contentType
+}
 
 module.exports = gotoOpts => {
   const goto = createGoto(gotoOpts)
@@ -23,10 +29,15 @@ module.exports = gotoOpts => {
     const [screenshotOpts, response] = await goto(page, url, opts)
     debug('goto', { duration: prettyMs(timeGoto()) })
 
-    if (codeScheme && response && isJSON(response.headers())) {
-      const timePretty = timeSpan()
-      await pReflect(pretty(page, response, { codeScheme, ...opts }))
-      debug('pretty', { duration: prettyMs(timePretty()) })
+    if (codeScheme && response) {
+      const headers = response.headers()
+      const contentType = getContentType(headers)
+
+      if (PRETTY_CONTENT_TYPES.includes(contentType)) {
+        const timePretty = timeSpan()
+        await pReflect(pretty(page, response, { codeScheme, contentType, ...opts }))
+        debug('pretty', { duration: prettyMs(timePretty()) })
+      }
     }
 
     const timeScreenshot = timeSpan()