+++ packages/devices/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.2.0](https://github.com/kikobeats/browserless/tree/master/packages/devices/compare/v7.1.7...v7.2.0) (2020-05-09)
+
+
+### Features
+
+* add `waitUntil: 'auto'` ([#171](https://github.com/kikobeats/browserless/tree/master/packages/devices/issues/171)) ([2116b67](https://github.com/kikobeats/browserless/tree/master/packages/devices/commit/2116b67ccddf210b76594fa71b053ee5c4b7e28c))
+
+
+
+
+
+# [7.0.0](https://github.com/kikobeats/browserless/tree/master/packages/devices/compare/v6.14.4...v7.0.0) (2020-04-19)
+
+
+### Features
+
+* use object map interface ([#165](https://github.com/kikobeats/browserless/tree/master/packages/devices/issues/165)) ([e78aac1](https://github.com/kikobeats/browserless/tree/master/packages/devices/commit/e78aac1335d04655b10e5dc4cb45d6e928aae89d))
+
+
+### BREAKING CHANGES
+
+* It needs puppeteer 3.x
+
+* fix: linter
+
+
+
+
+
+## [6.11.7](https://github.com/kikobeats/browserless/tree/master/packages/devices/compare/v6.11.6...v6.11.7) (2020-03-05)
+
+**Note:** Version bump only for package @browserless/devices
+
+
+
+
+
 ## [6.8.1](https://github.com/kikobeats/browserless/tree/master/packages/devices/compare/v6.8.0...v6.8.1) (2020-02-23)
 
 **Note:** Version bump only for package @browserless/devices