+++ packages/test/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.2.0](https://github.com/kikobeats/browserless/tree/master/packages/test/compare/v7.1.7...v7.2.0) (2020-05-09)
+
+
+### Features
+
+* add `waitUntil: 'auto'` ([#171](https://github.com/kikobeats/browserless/tree/master/packages/test/issues/171)) ([2116b67](https://github.com/kikobeats/browserless/tree/master/packages/test/commit/2116b67ccddf210b76594fa71b053ee5c4b7e28c))
+
+
+
+
+
+## [7.1.7](https://github.com/kikobeats/browserless/tree/master/packages/test/compare/v7.1.6...v7.1.7) (2020-05-09)
+
+**Note:** Version bump only for package @browserless/test
+
+
+
+
+
+## [7.1.4](https://github.com/kikobeats/browserless/tree/master/packages/test/compare/v7.1.3...v7.1.4) (2020-05-04)
+
+**Note:** Version bump only for package @browserless/test
+
+
+
+
+
+## [7.1.2](https://github.com/kikobeats/browserless/tree/master/packages/test/compare/v7.1.1...v7.1.2) (2020-04-22)
+
+**Note:** Version bump only for package @browserless/test
+
+
+
+
+
+# [7.0.0](https://github.com/kikobeats/browserless/tree/master/packages/test/compare/v6.14.4...v7.0.0) (2020-04-19)
+
+**Note:** Version bump only for package @browserless/test
+
+
+
+
+
 # [6.5.0](https://github.com/kikobeats/browserless/tree/master/packages/test/compare/v6.4.4...v6.5.0) (2020-02-12)
 
 **Note:** Version bump only for package @browserless/test